+++ css/info.css
@@ -150,3 +150,8 @@ a {
 }
 
 /* thẻ hr col-7 sp */
+.col-sm-10 {
+    padding: 0 !important;
+
+
+}
