+++ home/index.php
@@ -9,16 +9,11 @@
 
   <link rel="preconnect" href="https://fonts.googleapis.com" />
   <link rel="preconnect" href="https://fonts.gstatic.com" crossorigin />
-  <link
-    href="https://fonts.googleapis.com/css2?family=Josefin+Sans:wght@100&family=Lobster&family=Taviraj:wght@100;300&display=swap"
-    rel="stylesheet" />
+  <link href="https://fonts.googleapis.com/css2?family=Josefin+Sans:wght@100&family=Lobster&family=Taviraj:wght@100;300&display=swap" rel="stylesheet" />
 
-  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet"
-    integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous" />
+  <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-T3c6CoIi6uLrA9TneNEoa7RxnatzjcDSCmG1MXxSR1GAsXEV/Dwwykc2MPK8M2HN" crossorigin="anonymous" />
 
-  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css"
-    integrity="sha512-DTOQO9RWCH3ppGqcWaEA1BIZOC6xxalwEsw9c2QQeAIftl+Vegovlnee1c9QX4TctnWMn13TZye+giMm8e2LwA=="
-    crossorigin="anonymous" referrerpolicy="no-referrer" />
+  <link rel="stylesheet" href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.5.1/css/all.min.css" integrity="sha512-DTOQO9RWCH3ppGqcWaEA1BIZOC6xxalwEsw9c2QQeAIftl+Vegovlnee1c9QX4TctnWMn13TZye+giMm8e2LwA==" crossorigin="anonymous" referrerpolicy="no-referrer" />
 
   <style>
     body {
@@ -37,7 +32,7 @@
           <div class="col-lg-6 align-self-center">
             <p class="moto-text_system_2 mb-0">
               <span class="d-flex justify-content-center align-items-center">
-                <i class="fa-solid fa-phone fa-sm"></i></span>&nbsp;&nbsp;<a href="#" style="text-decoration: none;" >
+                <i class="fa-solid fa-phone fa-sm"></i></span>&nbsp;&nbsp;<a href="#" style="text-decoration: none;">
                 +84 395.399.892</a>&nbsp;Từ 9:00 đến 19:00 tất cả các ngày trong tuần
             </p>
           </div>
@@ -48,8 +43,7 @@
                   <div store-search-widget="" bis_skin_checked="1">
                     <form action="" method="GET" accept-charset="utf-8" class="search-form form-inline my-2 my-lg-0">
                       <label class="search-form_label">
-                        <input type="text" name="keyword" autocomplete="off" value=""
-                          class="search-form_input form-control" />
+                        <input type="text" name="keyword" autocomplete="off" value="" class="search-form_input form-control" />
                         <i class="fa-solid fa-magnifying-glass fa-sm search-icon"></i>
                       </label>
                       <span class="search-form_submit fa-search search-submit-icon"></span>
@@ -120,10 +114,10 @@
                         <script>
                           function logout() {
                             // Xử lý đăng xuất ở phía client (bằng JavaScript)
-                            window.location.href = '../login,logout,forgot/logout.php'; // Chuyển hướng đến tệp logout.php để xử lý đăng xuất
+                            window.location.href = '../Security/logout.php'; // Chuyển hướng đến tệp logout.php để xử lý đăng xuất
 
                             // Bạn có thể thêm dòng sau để chuyển hướng người dùng trở lại trang đăng nhập ngay sau khi đăng xuất
-                            // window.location.href = '../login,logout,forgot/login.php'; // Thay thế 'login.php' bằng trang bạn muốn chuyển hướng đến
+                            // window.location.href = '../Security/login.php'; // Thay thế 'login.php' bằng trang bạn muốn chuyển hướng đến
                           }
                         </script>
                       </div>
@@ -151,16 +145,13 @@
                     <div class="col-xs-12 d-flex align-items-center justify-content-center">
                       <div class="" data-widget="image">
                         <a href="" data-action="home_page" class="image-link moto-link">
-                          <img class="logo-link"
-                            src="https://demo.try-builder.com/site/03/00s/1w/0300s1wxnw55tckx/mt-content/uploads/2018/04/mt-1422_header_logo1.png"
-                            alt="">
+                          <img class="logo-link" src="https://demo.try-builder.com/site/03/00s/1w/0300s1wxnw55tckx/mt-content/uploads/2018/04/mt-1422_header_logo1.png" alt="">
                         </a>
                       </div>
                       <div class="">
                         <div class="" bis_skin_checked="1">
                           <p class="moto-text_system_1 d-flex align-items-center mb-0">
-                            <a data-action="home_page" data-anchor="" class="moto-link home_page logo-link" href=""
-                              bis_skin_checked="1">
+                            <a data-action="home_page" data-anchor="" class="moto-link home_page logo-link" href="" bis_skin_checked="1">
                               <span class="image-replacement"></span>
                               AZALEA
                             </a>
@@ -174,28 +165,23 @@
                       <a href="#" class="menu-toggle-btn"><i class="menu-toggle-btn-icon fa fa-bars"></i></a>
                       <ul class="menu-list menu-list_horizontal header m-0">
                         <li class="moto-widget-menu-item">
-                          <a href="#" data-action="home_page"
-                            class="menu-link menu-link-level-1 menu-link-active moto-link">TRANG CHỦ</a>
+                          <a href="#" data-action="home_page" class="menu-link menu-link-level-1 menu-link-active moto-link">TRANG CHỦ</a>
                         </li>
                         <li class="moto-widget-menu-item">
                           <a href="#" data-action="page" class="menu-link menu-link-level-1 moto-link">ABOUT US</a>
                         </li>
                         <li class="moto-widget-menu-item menu-item-has-submenu">
-                          <a href="../site/shop.php" data-action="store.category"
-                            class="menu-link menu-link-level-1 menu-link-submenu moto-link">SHOP
+                          <a href="../site/shop.php" data-action="store.category" class="menu-link menu-link-level-1 menu-link-submenu moto-link">SHOP
                             <i class="fa-solid fa-angle-down"></i></a>
                           <ul class="menu-sublist">
                             <li class="moto-widget-menu-item">
-                              <a href="../site/dress.php" data-action="store.category"
-                                class="menu-link menu-link-level-2 moto-link">VÁY CƯỚI</a>
+                              <a href="../site/dress.php" data-action="store.category" class="menu-link menu-link-level-2 moto-link">VÁY CƯỚI</a>
                             </li>
                             <li class="moto-widget-menu-item">
-                              <a href="../site/accessories.php" data-action="store.category"
-                                class="menu-link menu-link-level-2 moto-link">PHỤ KIỆN</a>
+                              <a href="../site/accessories.php" data-action="store.category" class="menu-link menu-link-level-2 moto-link">PHỤ KIỆN</a>
                             </li>
                             <li class="moto-widget-menu-item">
-                              <a href="collections.php" data-action="store.category"
-                                class="menu-link menu-link-level-2 moto-link">BỘ SƯU TẬP ĐỘC QUYỀN</a>
+                              <a href="collections.php" data-action="store.category" class="menu-link menu-link-level-2 moto-link">BỘ SƯU TẬP ĐỘC QUYỀN</a>
                             </li>
                           </ul>
                         </li>
@@ -225,10 +211,8 @@
                 </div>
               </div>
               <div class="container-fluid p-0">
-                <button class="btn btn-danger" data-widget-id="wid_1524136851_x7uuibamn" data-widget="button"
-                  bis_skin_checked="1">
-                  <a href="store/category/shop/" data-action="store.category"
-                    class="moto-widget-button-link moto-size-medium moto-link d-flex" bis_skin_checked="1">
+                <button class="btn btn-danger" data-widget-id="wid_1524136851_x7uuibamn" data-widget="button" bis_skin_checked="1">
+                  <a href="store/category/shop/" data-action="store.category" class="moto-widget-button-link moto-size-medium moto-link d-flex" bis_skin_checked="1">
                     <span class="fa moto-widget-theme-icon"></span>
                     <span class="moto-widget-button-divider"></span>
                     <span class="moto-widget-button-label text-white">XEM NGAY!</span>
@@ -248,9 +232,7 @@
         <div class="col-4 p-4">
           <a href="../site/dress.php" class="card-link">
             <div class="card" style="width: 100%">
-              <img
-                src="https://lavitta.vn/cdn/shop/files/z5009401811761_77bf56a6cf6eba0dbfd6a2fcc66288d7_600x.jpg?v=1703525193"
-                class="card-img-top" alt="mau1" />
+              <img src="https://lavitta.vn/cdn/shop/files/z5009401811761_77bf56a6cf6eba0dbfd6a2fcc66288d7_600x.jpg?v=1703525193" class="card-img-top" alt="mau1" />
               <div class="overlay">
                 <div class="content">
                   <p class="text-black">VÁY CƯỚI</p>
@@ -264,9 +246,7 @@
         <div class="col-4 p-4">
           <a href="../site/accessories.php" class="card-link">
             <div class="card" style="width: 100%">
-              <img
-                src="https://lavitta.vn/cdn/shop/files/z5009401802593_ac3db59bda6187bff17af1b8bbcfbbc5_600x.jpg?v=1703525097"
-                class="card-img-top" alt="mau2" />
+              <img src="https://lavitta.vn/cdn/shop/files/z5009401802593_ac3db59bda6187bff17af1b8bbcfbbc5_600x.jpg?v=1703525097" class="card-img-top" alt="mau2" />
               <div class="overlay">
                 <div class="content">
                   <p class="text-black">PHỤ KIỆN</p>
@@ -280,9 +260,7 @@
         <div class="col-4 p-4">
           <a href="../site/collections.php" class="card-link">
             <div class="card" style="width: 100%">
-              <img
-                src="https://lavitta.vn/cdn/shop/files/z5009401283292_9fe878edd75c243caa8d1b1ed6927490_600x.jpg?v=1703524257"
-                class="card-img-top" alt="mau3" />
+              <img src="https://lavitta.vn/cdn/shop/files/z5009401283292_9fe878edd75c243caa8d1b1ed6927490_600x.jpg?v=1703524257" class="card-img-top" alt="mau3" />
               <div class="overlay">
                 <div class="content">
                   <p class="text-black">BỘ SƯU TẬP ĐỘC QUYỀN</p>
@@ -309,10 +287,8 @@
               <p class="moto-text_system_6">FOR YOUR</p>
               <p class="moto-text_system_3">WEDDING</p>
               <div class="container-fluid p-0">
-                <button class="btn btn-danger" data-widget-id="wid_1524136851_x7uuibamn" data-widget="button"
-                  bis_skin_checked="1">
-                  <a href="store/category/shop/" data-action="store.category"
-                    class="moto-widget-button-link moto-size-medium moto-link d-flex" bis_skin_checked="1">
+                <button class="btn btn-danger" data-widget-id="wid_1524136851_x7uuibamn" data-widget="button" bis_skin_checked="1">
+                  <a href="store/category/shop/" data-action="store.category" class="moto-widget-button-link moto-size-medium moto-link d-flex" bis_skin_checked="1">
                     <span class="fa moto-widget-theme-icon"></span>
                     <span class="moto-widget-button-divider"></span>
                     <span class="moto-widget-button-label text-white">XEM NGAY!</span>
@@ -329,25 +305,22 @@
   <div class="container py-5">
     <div class="row">
       <h1 class="text-center text-white">
-        <img data-src="" src="https://demo.try-builder.commt-content/uploads/2018/04/mt-1422_header_logo1.png" class=""
-          data-id="1120" title="" alt="" />
+        <img data-src="" src="https://demo.try-builder.commt-content/uploads/2018/04/mt-1422_header_logo1.png" class="" data-id="1120" title="" alt="" />
       </h1>
 
       <h1 class="text-center text-white mb-4">OUR PRODUCTS</h1>
       <?php
       include 'C:/xampp/php/BAITAP/BTLPHP/db/dbconnect.php';
-
-      $sql = "SELECT * FROM `image_demo_vaycuoi`";
+      $sql = "SELECT tb_image.id_sp, tb_image.name_sp, tb_image.price, tb_image.image_sp, tb_image.code_sp, MAX(price_sp.price) AS max_price FROM tb_image JOIN price_sp ON tb_image.code_sp = price_sp.code_sp GROUP BY tb_image.code_sp;";
       $result = $conn->query($sql);
 
       if ($result->num_rows > 0) {
         // Duyệt qua từng dòng dữ liệu
         while ($row = $result->fetch_assoc()) {
-          ?>
+      ?>
           <div class="col-sm-3">
-            <div class="card sp" style="width: 100%" onmouseover="addHoverEffect(this)" onmouseout="removeHoverEffect(this)"
-              onclick="redirectToAnotherPage()">
-              <img class="h-100" src="data:image/jpeg;base64,<?= base64_encode($row['image_sp_demo']) ?>" alt="" />
+            <div class="card sp" style="width: 100%" onmouseover="addHoverEffect(this)" onmouseout="removeHoverEffect(this)" onclick="redirectToAnotherPage()">
+              <img class="h-100" src="data:image/jpeg;base64,<?= base64_encode($row['image_sp']) ?>" alt="" />
               <a href="#" class="eye-link">
                 <i class="fa-regular fa-eye"></i>
               </a>
@@ -355,9 +328,9 @@
 
             <div class="d-flex justify-content-center pt-3">
               <p class="m-0 text-center">
-                <?= $row['name_sp_demo'] ?><br>
+                <?= $row['name_sp'] ?><br>
                 <?= '<span style="color: #dc3545;">' . $row['price'] . '</span>' ?><br>
-                <!-- <?= $row['status'] ?> -->
+                <!-- <?= $row['code_sp'] ?> -->
               </p>
             </div>
 
@@ -367,7 +340,7 @@
               </button>
             </div>
           </div>
-          <?php
+      <?php
         }
       } else {
         echo "Không có dữ liệu";
@@ -383,8 +356,7 @@
   <div class="container py-5">
     <div class="row py-5">
       <div class="col-sm-6">
-        <img class="w-100 rounded"
-          src="https://mymedia.vn/wp-content/uploads/2022/08/chup-anh-cuoi-co-dau-mot-minh-2.jpg" alt="" />
+        <img class="w-100 rounded" src="https://mymedia.vn/wp-content/uploads/2022/08/chup-anh-cuoi-co-dau-mot-minh-2.jpg" alt="" />
       </div>
       <div class="col-sm-6 d-flex align-items-center">
         <div class="">
@@ -407,8 +379,7 @@
         </div>
       </div>
       <div class="col-sm-6">
-        <img class="w-100 rounded"
-          src="https://mimosawedding.net/wp-content/uploads/2021/10/chup-anh-cuoi-co-dau-mot-minh-20.jpg" alt="" />
+        <img class="w-100 rounded" src="https://mimosawedding.net/wp-content/uploads/2021/10/chup-anh-cuoi-co-dau-mot-minh-20.jpg" alt="" />
       </div>
     </div>
   </div>
@@ -417,46 +388,30 @@
     <div class="">
       <div class="container-fluid" bis_skin_checked="1">
         <div class="row" data-container="container" bis_skin_checked="1">
-          <div class="col-sm-12" data-widget="row.column" data-container="container" data-spacing="aaaa"
-            data-bg-position="left top" bis_skin_checked="1">
-            <div class="" data-grid-type="sm" data-widget="row" data-spacing="aama" data-bg-position="left top"
-              data-draggable-disabled="" bis_skin_checked="1">
+          <div class="col-sm-12" data-widget="row.column" data-container="container" data-spacing="aaaa" data-bg-position="left top" bis_skin_checked="1">
+            <div class="" data-grid-type="sm" data-widget="row" data-spacing="aama" data-bg-position="left top" data-draggable-disabled="" bis_skin_checked="1">
               <div class="container-fluid" bis_skin_checked="1">
                 <div class="row" data-container="container" bis_skin_checked="1">
-                  <div class="col-sm-4" data-widget="row.column" data-container="container" data-spacing="aaaa"
-                    data-bg-position="left top" bis_skin_checked="1"></div>
-                  <div class="col-sm-4" data-widget="row.column" data-container="container" data-spacing="aaaa"
-                    data-bg-position="left top" bis_skin_checked="1">
-                    <div class="" data-grid-type="xs" data-widget="row" data-spacing="aaaa" data-bg-position="left top"
-                      bis_skin_checked="1">
+                  <div class="col-sm-4" data-widget="row.column" data-container="container" data-spacing="aaaa" data-bg-position="left top" bis_skin_checked="1"></div>
+                  <div class="col-sm-4" data-widget="row.column" data-container="container" data-spacing="aaaa" data-bg-position="left top" bis_skin_checked="1">
+                    <div class="" data-grid-type="xs" data-widget="row" data-spacing="aaaa" data-bg-position="left top" bis_skin_checked="1">
                       <div class="container-fluid" bis_skin_checked="1">
                         <div class="row" data-container="container" bis_skin_checked="1">
-                          <div class="col-xs-2" data-widget="row.column" data-container="container" data-spacing="aaaa"
-                            data-bg-position="left top" bis_skin_checked="1"></div>
-                          <div class="col-xs-9 py-3 d-flex justify-content-center" data-widget="row.column"
-                            data-container="container" data-spacing="aaaa" data-bg-position="left top"
-                            bis_skin_checked="1">
-                            <div class="" data-grid-type="xs" data-widget="row" data-spacing="sasa"
-                              data-bg-position="left top" data-draggable-disabled="" bis_skin_checked="1">
+                          <div class="col-xs-2" data-widget="row.column" data-container="container" data-spacing="aaaa" data-bg-position="left top" bis_skin_checked="1"></div>
+                          <div class="col-xs-9 py-3 d-flex justify-content-center" data-widget="row.column" data-container="container" data-spacing="aaaa" data-bg-position="left top" bis_skin_checked="1">
+                            <div class="" data-grid-type="xs" data-widget="row" data-spacing="sasa" data-bg-position="left top" data-draggable-disabled="" bis_skin_checked="1">
                               <div class="container-fluid" bis_skin_checked="1">
                                 <div class="row align-items-center" data-container="container" bis_skin_checked="1">
-                                  <div class="col-xs-12 d-flex align-items-center justify-content-center"
-                                    data-widget="row.column" data-container="container" data-spacing="aaaa"
-                                    data-bg-position="left top" bis_skin_checked="1">
+                                  <div class="col-xs-12 d-flex align-items-center justify-content-center" data-widget="row.column" data-container="container" data-spacing="aaaa" data-bg-position="left top" bis_skin_checked="1">
                                     <div class="" data-widget="image" bis_skin_checked="1">
-                                      <a href="" data-action="home_page" class="image-link moto-link"
-                                        bis_skin_checked="1">
-                                        <img class="logo-link"
-                                          src="https://demo.try-builder.com/site/03/00s/1w/0300s1wxnw55tckx/mt-content/uploads/2018/04/mt-1422_header_logo1.png"
-                                          alt="">
+                                      <a href="" data-action="home_page" class="image-link moto-link" bis_skin_checked="1">
+                                        <img class="logo-link" src="https://demo.try-builder.com/site/03/00s/1w/0300s1wxnw55tckx/mt-content/uploads/2018/04/mt-1422_header_logo1.png" alt="">
                                       </a>
                                     </div>
-                                    <div class="" data-widget="text" data-preset="default" data-spacing="aaaa"
-                                      data-animation="" data-draggable-disabled="" bis_skin_checked="1">
+                                    <div class="" data-widget="text" data-preset="default" data-spacing="aaaa" data-animation="" data-draggable-disabled="" bis_skin_checked="1">
                                       <div class="" bis_skin_checked="1">
                                         <p class="moto-text_system_1 d-flex align-items-center mb-0">
-                                          <a data-action="home_page" data-anchor=""
-                                            class="moto-link home_page logo-link" href="" bis_skin_checked="1">
+                                          <a data-action="home_page" data-anchor="" class="moto-link home_page logo-link" href="" bis_skin_checked="1">
                                             <span class="image-replacement"></span>
                                             AZALEA
                                           </a>
@@ -468,84 +423,59 @@
                               </div>
                             </div>
                           </div>
-                          <div class="col-xs-1" data-widget="row.column" data-container="container" data-spacing="aaaa"
-                            data-bg-position="left top" bis_skin_checked="1"></div>
+                          <div class="col-xs-1" data-widget="row.column" data-container="container" data-spacing="aaaa" data-bg-position="left top" bis_skin_checked="1"></div>
                         </div>
                       </div>
                     </div>
                   </div>
-                  <div class="col-sm-4" data-widget="row.column" data-container="container" data-spacing="aaaa"
-                    data-bg-position="left top" bis_skin_checked="1"></div>
+                  <div class="col-sm-4" data-widget="row.column" data-container="container" data-spacing="aaaa" data-bg-position="left top" bis_skin_checked="1"></div>
                 </div>
               </div>
             </div>
-            <div data-widget-id="wid_1524128091_05kg60ys0" class="" data-preset="default" data-widget="menu"
-              bis_skin_checked="1">
+            <div data-widget-id="wid_1524128091_05kg60ys0" class="" data-preset="default" data-widget="menu" bis_skin_checked="1">
               <a href="#" class="menu-toggle-btn"><i class=" fa fa-bars"></i></a>
               <ul class="menu-list menu-list_horizontal footer">
                 <li class="moto-widget-menu-item">
-                  <a href="index.php" data-action="home_page"
-                    class="menu-link menu-link-level-1 menu-link-active moto-link" bis_skin_checked="1">TRANG CHỦ</a>
+                  <a href="index.php" data-action="home_page" class="menu-link menu-link-level-1 menu-link-active moto-link" bis_skin_checked="1">TRANG CHỦ</a>
                 </li>
                 <li class="moto-widget-menu-item">
                   <a href="" data-action="page" class="menu-link menu-link-level-1 moto-link" bis_skin_checked="1">ABOUT
                     US</a>
                 </li>
                 <li class="moto-widget-menu-item">
-                  <a href="" data-action="blog.index" class="menu-link menu-link-level-1 moto-link"
-                    bis_skin_checked="1">BLOG</a>
+                  <a href="" data-action="blog.index" class="menu-link menu-link-level-1 moto-link" bis_skin_checked="1">BLOG</a>
                 </li>
                 <li class="moto-widget-menu-item">
-                  <a href="../site/shop.php" data-action="store.category" class="menu-link menu-link-level-1 moto-link"
-                    bis_skin_checked="1">SHOP</a>
+                  <a href="../site/shop.php" data-action="store.category" class="menu-link menu-link-level-1 moto-link" bis_skin_checked="1">SHOP</a>
                 </li>
                 <li class="moto-widget-menu-item">
-                  <a href="" data-action="page" class="menu-link menu-link-level-1 moto-link"
-                    bis_skin_checked="1">LIÊN HỆ</a>
+                  <a href="" data-action="page" class="menu-link menu-link-level-1 moto-link" bis_skin_checked="1">LIÊN HỆ</a>
                 </li>
               </ul>
             </div>
-            <div class="" data-widget="text" data-preset="default" data-spacing="saaa" data-animation=""
-              data-draggable-disabled="" bis_skin_checked="1">
+            <div class="" data-widget="text" data-preset="default" data-spacing="saaa" data-animation="" data-draggable-disabled="" bis_skin_checked="1">
               <div class="" bis_skin_checked="1">
                 <p class="moto-text_system_8" style="text-align: center">
-                  <a target="_self" data-action="url" class="moto-link payment" href="#"><span class="fa"><i
-                        class="fa-brands fa-cc-visa fa-xl payment"></i></span></a>&nbsp;<a target="_self"
-                    data-action="url" class="moto-link" href="#"><span class="fa"><i
-                        class="fa-brands fa-cc-mastercard fa-xl payment"></i></span></a>&nbsp;<a target="_self"
-                    data-action="url" class="moto-link" href="#"><span class="fa"><i
-                        class="fa-brands fa-cc-discover fa-xl payment"></i></span></a>&nbsp;<a target="_self"
-                    data-action="url" class="moto-link" href="#"><span class="fa"><i
-                        class="fa-brands fa-cc-apple-pay fa-xl payment"></i></span></a>&nbsp;<a target="_self"
-                    data-action="url" class="moto-link" href="#"><span class="fa"><i
-                        class="fa-brands fa-cc-paypal fa-xl payment"></i></span></a>&nbsp;<a target="_self"
-                    data-action="url" class="moto-link" href="#"><span class="fa"><i
-                        class="fa-brands fa-cc-stripe fa-xl payment"></i></span></a>
+                  <a target="_self" data-action="url" class="moto-link payment" href="#"><span class="fa"><i class="fa-brands fa-cc-visa fa-xl payment"></i></span></a>&nbsp;<a target="_self" data-action="url" class="moto-link" href="#"><span class="fa"><i class="fa-brands fa-cc-mastercard fa-xl payment"></i></span></a>&nbsp;<a target="_self" data-action="url" class="moto-link" href="#"><span class="fa"><i class="fa-brands fa-cc-discover fa-xl payment"></i></span></a>&nbsp;<a target="_self" data-action="url" class="moto-link" href="#"><span class="fa"><i class="fa-brands fa-cc-apple-pay fa-xl payment"></i></span></a>&nbsp;<a target="_self" data-action="url" class="moto-link" href="#"><span class="fa"><i class="fa-brands fa-cc-paypal fa-xl payment"></i></span></a>&nbsp;<a target="_self" data-action="url" class="moto-link" href="#"><span class="fa"><i class="fa-brands fa-cc-stripe fa-xl payment"></i></span></a>
                 </p>
               </div>
             </div>
-            <div data-widget-id="wid_1524128807_2abosenqb" class="" data-widget="divider_horizontal" data-preset="3"
-              bis_skin_checked="1">
+            <div data-widget-id="wid_1524128807_2abosenqb" class="" data-widget="divider_horizontal" data-preset="3" bis_skin_checked="1">
               <hr class="moto-widget-divider-line" style="max-width: 100%; width: 100%" />
             </div>
-            <div class="" data-widget="text" data-preset="default" data-spacing="aaaa" data-animation=""
-              data-draggable-disabled="" bis_skin_checked="1">
+            <div class="" data-widget="text" data-preset="default" data-spacing="aaaa" data-animation="" data-draggable-disabled="" bis_skin_checked="1">
               <div class="moto-widget-text-content moto-widget-text-editable" bis_skin_checked="1">
                 <p class="moto-text_224" style="text-align: center">
                   © 2024 Azalea. All Right Reserved. Designed by
-                  <a target="_blank" rel="nofollow" data-action="url" class="moto-link author"
-                    href="http://github.com/hvuitsme" bis_skin_checked="1">Hvuitsme</a>.&nbsp;<a href=""
-                    data-action="page" data-id="27" class="moto-link policy" bis_skin_checked="1">Privacy Policy</a>
+                  <a target="_blank" rel="nofollow" data-action="url" class="moto-link author" href="http://github.com/hvuitsme" bis_skin_checked="1">Hvuitsme</a>.&nbsp;<a href="" data-action="page" data-id="27" class="moto-link policy" bis_skin_checked="1">Privacy Policy</a>
                 </p>
               </div>
             </div>
-            <div id="wid_1524129043_jzrillzxh" data-widget-id="wid_1524129043_jzrillzxh" class=""
-              data-widget="social_links_extended" data-preset="default" bis_skin_checked="1">
+            <div id="wid_1524129043_jzrillzxh" data-widget-id="wid_1524129043_jzrillzxh" class="" data-widget="social_links_extended" data-preset="default" bis_skin_checked="1">
               <ul class="social-links-extended__list">
                 <li class="social-links-extended__item ">
                   <a href="#" class="moto-widget-social-links-extended__link" target="_blank">
-                    <span class="moto-widget-social-links-extended__icon fa"><i
-                        class="fa-brands fa-square-facebook fa-xl"></i></span>
+                    <span class="moto-widget-social-links-extended__icon fa"><i class="fa-brands fa-square-facebook fa-xl"></i></span>
                   </a>
                 </li>
                 <li class="social-links-extended__item ">
@@ -555,14 +485,12 @@
                 </li>
                 <li class="social-links-extended__item ">
                   <a href="#" class="" target="_blank">
-                    <span class="moto-widget-social-links-extended__icon fa"><i
-                        class="fa-brands fa-linkedin fa-xl"></i></span>
+                    <span class="moto-widget-social-links-extended__icon fa"><i class="fa-brands fa-linkedin fa-xl"></i></span>
                   </a>
                 </li>
                 <li class="social-links-extended__item ">
                   <a href="#" class="moto-widget-social-links-extended__link" target="_blank">
-                    <span class="moto-widget-social-links-extended__icon fa"><i
-                        class="fa-brands fa-instagram fa-xl"></i></span>
+                    <span class="moto-widget-social-links-extended__icon fa"><i class="fa-brands fa-instagram fa-xl"></i></span>
                   </a>
                 </li>
               </ul>
@@ -577,7 +505,7 @@
   <!-- js script -->
 
   <script>
-    $(document).ready(function () {
+    $(document).ready(function() {
       // Kiểm tra kích thước màn hình và thêm/xóa lớp CSS
       function checkScreenSize() {
         var screenSize =
@@ -594,7 +522,7 @@
       }
 
       // Gọi hàm kiểm tra khi trang web tải và khi thay đổi kích thước màn hình
-      $(window).on("load resize", function () {
+      $(window).on("load resize", function() {
         checkScreenSize();
       });
     });
@@ -619,9 +547,7 @@
 
   <script src="https://code.jquery.com/jquery-3.6.4.min.js"></script>
 
-  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js"
-    integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL"
-    crossorigin="anonymous"></script>
+  <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.3.2/dist/js/bootstrap.bundle.min.js" integrity="sha384-C6RzsynM9kWDrMNeT87bh95OGNyZPhcTNXj1NW7RuBCsyN/o0jlpcV8Qyq46cDfL" crossorigin="anonymous"></script>
 </body>
 
 </html>
